requirements.txt
boto3==1.34.0
pandas==2.1.4
python-dotenv==1.0.0